Customizing & Using UserRegistrationFormType
Cool. It set the data_class to User and even looked at the properties on that class and pre-filled the fields! Let's see: we don't want roles or twitterUsername for registration. And, firstName is something that I won't include either - the current form has just these two fields: email and password.
Ok: step 2: go the controller and create the form! And, yes! I get to remove a 'TODO' in my code - that never happens! Use the normal $ form = $ this->createForm() and pass this UserRegistrationFormType::class. But don't pass a second argument: we want the form to create a new User object.

Then, add $ form->handleRequest($ request) and, for the if, use ->isSubmitted() && $ form->isValid().
Beautiful, boring, normal, code. And now that we're using the form system, instead of creating the User object like chumps, say $ user = $ form->getData(). I'll add some inline documentation so that PhpStorm knows what this variable is. Oh, and we don't need to set the email directly anymore: the form will do that! And I'll remove my firstName hack: we'll fix that in a minute.
Down below when we render the template, pass a new registrationForm variable set to $ form->createView().
About the password: we do need to encode the password. But now, the plain text password will be stored on $ user->getPassword(). Hmm. That is a little weird: the form system is setting the plaintext password on the password field. And then, a moment later, we're encoding that and setting it back on that same property! We're going to change this in a few minutes - but, it should work.