The second solution is... a bit more interesting: we can mark the field to not be 'mapped'. Check it out: pass null as the second argument to add() so it continues guessing the field type for now. Then, pass a new option: mapped set to false.
That changes everything. This tells the form system that we do want to have this plainPassword field on our form, but that it should not get or set its data back onto the User object. It means that we no longer need getPlainPassword() and setPlainPassword() methods!

Accessing Unmapped Fields
Woo! Except... wait, if the form doesn't set this data onto the User object... how the heck can we access that data? After all, when we call $ form->getData(), it gives us the User object. Where will that plainPassword data live?

Then move over, refresh.  Yes! It hits our dump and die and there is our plain password!

This uncovers a really neat thing about the form system. When you call $ this->createForm(), it creates a Form object that represents the whole form. But also, each individual field is also represented as its own Form object, and it's a child of that top-level form. Yep, $ form['plainPassword'] gives us a Form object that knows everything about this one field. When we call ->getData() on it, yep! That's the value for this one field.

This is a super nice solution for situations where you need to add a field to your form, but it doesn't map cleanly to a property on your entity.