Change this to PasswordType::class. This won't change how the field behaves, only how it's rendered. Yep! A proper <input type='password'> field.
Next: time to add validation! Which, hmm, is going to be a bit interesting. First, we need to validate that the user is unique in the database. And second, for the first time, we need to add validation to a form field where there is no corresponding property on our class.