The User class has an agreedTermsAt property that expects a DateTime object. But, our form has an agreeTerms field that, on submit, will give us a true/false boolean value. How can we make these work together? As I so often like to say: there are two options.
First, we could be clever! There is no agreeTerms property on User.
But, we could create a setAgreeTerms() method on User like:
public function setAgreeTerms(bool $ agree){
  $ this->agreedTermsAt = $ agree ? new \DateTime : null;
}
We would also need to create a getAgreeTerms() like:
public function getAgreeTerms():bool {
  return $ this->agreedTermsAt !== null;
}

This is a fine solution. But, this is also a good example of how the form system can start to make your life harder instead of easier.