There's just one problem left with our registration form - it looks terrible! It does not look like our original form, which was styled pretty nicely. One of the trickiest things to do with the form system is to style, or theme your forms. The system is super powerful: we just need to unlock its potential!
Adding Attributes to the Tag

Here's the goal: make our form render the same markup that we had before. Let's start with something simple: the form tag had a class called form-signin. Google for 'Symfony form function reference'. Hey, we know this page! It lists all of the functions that we can call to render each part of our form! And it will give us a clue about how we can customize each part.

For example, the first argument to form_start() is called view. When you see 'view' on this page, it's referring to your form variable. The really important argument is the second one: variables. We saw this before: almost every function has this mysterious variables argument. This is an array of, literally, Twig variables! They're used to render each part of the form.

For example, there is apparently a variable called method that you can set to control the method attribute on the form. But, it's not as simple as: every variable becomes an attribute. For example, we can't just pass a class variable to add a class attribute.