The buildView() Form Class Method
To fix this, at the bottom of UserSelectTextType, go to the Code -> Generate menu, or command+N on a Mac, select Override methods and choose buildView(). Oh, there's also a method called finishView() and its purpose is almost identical to buildView() - it's just called a bit later.

Here's what's going on: to render each field, Symfony creates a bunch of variables that are used in the form theme system. We already knew that: in register.html.twig we're overriding the attr variable. And in our form theme blocks, we use different variables to do our work.

And, of course, we know that, thanks to the profiler, we can see the exact view variables that exist for each field. But... where do these variables come from? For example, why does each field have a full_name variable? Who added that?

The answer is buildView(): Symfony calls this method on every field, and it is the place where these variables are created and can be changed.

We do that with this $ view variable, which is kind of a strange object. Start with $ attr = $ view->vars['attr'];. This $ view object has a public ->vars array property that holds all of the things that will eventually become the 'variables'. At this moment, the core form system has already set this variable up for us: it will either be equal to the attr option passed for this field, or an empty array.

Next: grab the class: if class is set on $ attr, use it, but add a space on the end. If there is no class yet, set this to be blank. Now, here's the key: let's always append js-user-autocomplete: that's the class we're using above. Call $ attr['class'] = to set the new class string back on.

Oh, and we also need to add the data-autocomplete-url attribute. Copy that from above and say $ attr['data-autocomplete-url'] equals the generated URL. Perfect! Finally, set all of this back onto the view object with $ view->vars['attr'] = $ attr.

Phew! We're done! Now that we're setting the attr variable directly, we don't need to set the option anymore. And the best part is that we know our attributes will be rendered no matter what the user passes to the attr option.

Let's try it! Move over, refresh and cool! Nice work team! The element still has the attributes we need.

Oh, and open the profiler for this form. Click on the author field and check out the View Variables. So cool! That's exactly what we set!

Next: the form component has a crazy powerful plugin system. Want to make some tweak to every form or even every field in your entire app? That's possible, and it's fun!